[Refactor Codes] activity_setting.xml 업데이트 된 뷰 구성 완료 설정 뷰, 개발자 정보, 버전 정보 뷰 업데이트
